ruuvitag/requirements.txt
paho-mqtt==1.4.0
python-dotenv==0.10.3
ruuvitag_sensor==0.13.0